+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/PlanWaitInstruction.cs
@@ -22,13 +22,13 @@ namespace Assistant.NINAPlugin.Sequencer {
         }
 
         public override Task Execute(IProgress<ApplicationStatus> progress, CancellationToken token) {
-            Logger.Info($"Scheduler: stopping guiding/tracking, then waiting for {Utils.FormatDateTimeFull(waitUntil)}");
+            TSLogger.Info($"stopping guiding/tracking, then waiting for {Utils.FormatDateTimeFull(waitUntil)}");
             SequenceCommands.StopGuiding(guiderMediator, token);
             SequenceCommands.SetTelescopeTracking(telescopeMediator, TrackingMode.Stopped, token);
 
             TimeSpan duration = ((DateTime)waitUntil) - DateTime.Now;
             CoreUtil.Wait(duration, token, progress).Wait(token);
-            Logger.Debug("Scheduler: done waiting");
+            TSLogger.Debug("done waiting");
 
             return Task.CompletedTask;
         }