+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/SequenceCommands.cs
@@ -1,4 +1,5 @@
-﻿using NINA.Core.Model;
+﻿using Assistant.NINAPlugin.Util;
+using NINA.Core.Model;
 using NINA.Core.Utility;
 using NINA.Equipment.Interfaces;
 using NINA.Equipment.Interfaces.Mediator;
@@ -12,41 +13,41 @@ namespace Assistant.NINAPlugin.Sequencer {
 
         public static void StopGuiding(IGuiderMediator guiderMediator, CancellationToken token) {
             if (guiderMediator.GetInfo().Connected) {
-                Logger.Info("Scheduler: stopping guiding");
+                TSLogger.Info("stopping guiding");
                 var stoppedGuiding = Task.Run(async () => {
                     await guiderMediator.StopGuiding(token);
                 });
             }
             else {
-                Logger.Warning("Scheduler: no guider connected, skipping StopGuiding");
+                TSLogger.Warning("no guider connected, skipping StopGuiding");
             }
         }
 
         public static void StartGuiding(IGuiderMediator guiderMediator, IProgress<ApplicationStatus> progress, CancellationToken token) {
             if (guiderMediator.GetInfo().Connected) {
-                Logger.Info("Scheduler: starting guiding");
+                TSLogger.Info("starting guiding");
                 var stoppedGuiding = Task.Run(async () => {
                     await guiderMediator.StartGuiding(false, progress, token);
                 });
             }
             else {
-                Logger.Warning("Scheduler: no guider connected, skipping StartGuiding");
+                TSLogger.Warning("no guider connected, skipping StartGuiding");
             }
         }
 
         public static void SetTelescopeTracking(ITelescopeMediator telescopeMediator, TrackingMode trackingMode, CancellationToken token) {
             if (telescopeMediator.GetInfo().Connected) {
-                Logger.Info($"Scheduler: set mount tracking to {trackingMode}");
+                TSLogger.Info($"set mount tracking to {trackingMode}");
                 telescopeMediator.SetTrackingMode(trackingMode);
             }
             else {
-                Logger.Warning("Scheduler: no mount connected, skipping SetTelescopeTracking");
+                TSLogger.Warning("no mount connected, skipping SetTelescopeTracking");
             }
         }
 
         public static async Task ParkTelescope(ITelescopeMediator telescopeMediator, IGuiderMediator guiderMediator, IProgress<ApplicationStatus> progress, CancellationToken token) {
             if (telescopeMediator.GetInfo().Connected) {
-                Logger.Info($"Scheduler: parking telescope");
+                TSLogger.Info($"parking telescope");
                 StopGuiding(guiderMediator, token);
                 await guiderMediator.StopGuiding(token);
                 if (!await telescopeMediator.ParkTelescope(progress, token)) {
@@ -54,32 +55,32 @@ namespace Assistant.NINAPlugin.Sequencer {
                 }
             }
             else {
-                Logger.Warning("Scheduler: no mount connected, skipping ParkTelescope");
+                TSLogger.Warning("no mount connected, skipping ParkTelescope");
             }
         }
 
         public static async Task UnparkTelescope(ITelescopeMediator telescopeMediator, IProgress<ApplicationStatus> progress, CancellationToken token) {
             if (telescopeMediator.GetInfo().Connected) {
-                Logger.Info($"Scheduler: unparking telescope");
+                TSLogger.Info($"unparking telescope");
                 bool success = await telescopeMediator.UnparkTelescope(progress, token);
                 if (!success) {
                     throw new SequenceEntityFailedException();
                 }
             }
             else {
-                Logger.Warning("Scheduler: no mount connected, skipping UnparkTelescope");
+                TSLogger.Warning("no mount connected, skipping UnparkTelescope");
             }
         }
 
         public static async Task ParkDome(IDomeMediator domeMediator, CancellationToken token) {
             if (domeMediator.GetInfo().Connected) {
-                Logger.Info($"Scheduler: parking dome");
+                TSLogger.Info($"parking dome");
                 if (!await domeMediator.Park(token)) {
                     throw new SequenceEntityFailedException();
                 }
             }
             else {
-                Logger.Warning("Scheduler: no dome connected, skipping ParkDome");
+                TSLogger.Warning("no dome connected, skipping ParkDome");
             }
         }